requirements.txt
pytest==7.4.3
Appium-Python-Client==3.1.0
